+++ dawg.h
@@ -18,7 +18,7 @@ class Dawg {
     Dawg();
 
     int process(int mode, uint32_t param);
-    int traverse(uint32_t ptr, int buf_ptr, uint32_t hash);
+    int traverse(uint16_t ptr, int buf_ptr, uint32_t hash);
     int traverse();
 
     int sort_results();
